Fractional reader progress: teleprompter-smooth strip
wordIndex is now fractional (word index + fraction of that word's
phonemes reached) from both /align modes, and the strip interpolates
its scroll position between word centers — the reader slides
continuously instead of stepping 2-4 whole words per update.

- scoring: _word_progress maps consumed tokens to fractional words;
  align_progress and locate_window return floats
- backend: positions stored as float64 bits; auto-finish arms when the
  last word is half consumed; peek caps/filter operate on floats
- peeks: cadence 0.33s -> 0.25s, window 2.5s -> 2.0s (faster
  transcription, still enough context for the phoneme model)
- frontend: WordStrip interpolates translateX between word centers by
  the fractional part; Home passes the float through directly

+++ backend/internal/inference/client.go
@@ -138,11 +138,12 @@ func (c *Client) Score(ctx context.Context, text, actualIPA string) (*ScoreRespo
 }
 
 // Align posts the reference text and IPA to /align (mode "prefix": actualIPA
-// is the cumulative transcript) and returns the last matched word index
-// (-1 = nothing matched yet). maxWord >= 0 hard-caps the answer — chunks
-// arrive every ~2.5 s, so the position can physically advance only so far
-// between them. Pass a negative maxWord for no cap.
-func (c *Client) Align(ctx context.Context, text, actualIPA string, maxWord int) (int, error) {
+// is the cumulative transcript) and returns the fractional reading progress
+// (word index + fraction of that word's phonemes reached; -1 = nothing
+// matched yet). maxWord >= 0 hard-caps the answer — chunks arrive every
+// ~2.5 s, so the position can physically advance only so far between them.
+// Pass a negative maxWord for no cap.
+func (c *Client) Align(ctx context.Context, text, actualIPA string, maxWord int) (float64, error) {
 	payload := map[string]any{
 		"text":      text,
 		"actualIpa": actualIPA,
@@ -157,7 +158,8 @@ func (c *Client) Align(ctx context.Context, text, actualIPA string, maxWord int)
 // AlignWindow is mode "window": actualIPA is a short trailing window, aligned
 // locally against the band of words starting at fromWord, never answering
 // beyond maxWord (a further match is always a duplicate-word coincidence).
-func (c *Client) AlignWindow(ctx context.Context, text, actualIPA string, fromWord, maxWord int) (int, error) {
+// Returns fractional progress like Align.
+func (c *Client) AlignWindow(ctx context.Context, text, actualIPA string, fromWord, maxWord int) (float64, error) {
 	return c.align(ctx, map[string]any{
 		"text":      text,
 		"actualIpa": actualIPA,
@@ -167,7 +169,7 @@ func (c *Client) AlignWindow(ctx context.Context, text, actualIPA string, fromWo
 	})
 }
 
-func (c *Client) align(ctx context.Context, payload map[string]any) (int, error) {
+func (c *Client) align(ctx context.Context, payload map[string]any) (float64, error) {
 	body, err := json.Marshal(payload)
 	if err != nil {
 		return -1, err
@@ -188,7 +190,7 @@ func (c *Client) align(ctx context.Context, payload map[string]any) (int, error)
 		return -1, fmt.Errorf("inference /align returned %s: %s", resp.Status, strings.TrimSpace(string(b)))
 	}
 	var out struct {
-		WordIndex int `json:"wordIndex"`
+		WordIndex float64 `json:"wordIndex"`
 	}
 	if err := json.NewDecoder(resp.Body).Decode(&out); err != nil {
 		return -1, fmt.Errorf("decoding /align response: %w", err)
